+++ apps/netconf/netconf_rpc.c
@@ -671,7 +671,7 @@ netconf_rpc_dispatch(clicon_handle h,
      * It may even be wrong if something else is done with the incoming message?
      */
     if ((username = clicon_username_get(h)) != NULL){
-	if ((xa = xml_new("username", NULL, xn, CX_ATTR)) == NULL)
+	if ((xa = xml_new("username", xn, CX_ATTR)) == NULL)
 	    goto done;
 	if (xml_value_set(xa, username) < 0)
 	    goto done;