+++ lib/clixon/clixon_xml.h
@@ -146,7 +146,6 @@ char     *xml_value(cxobj *xn);
 int       xml_value_set(cxobj *xn, char *val);
 int       xml_value_append(cxobj *xn, char *val);
 enum cxobj_type xml_type(cxobj *xn);
-int       xml_type_set(cxobj *xn, enum cxobj_type type);
 
 int       xml_child_nr(cxobj *xn);
 int       xml_child_nr_type(cxobj *xn, enum cxobj_type type);
@@ -160,8 +159,8 @@ cxobj    *xml_child_each(cxobj *xparent, cxobj *xprev,  enum cxobj_type type);
 int       xml_child_insert_pos(cxobj *x, cxobj *xc, int i);
 int       xml_childvec_set(cxobj *x, int len);
 cxobj   **xml_childvec_get(cxobj *x);
-cxobj    *xml_new(char *name, char *prefix, cxobj *xn_parent, enum cxobj_type type);
-
+cxobj    *xml_new(char *name, cxobj *xn_parent, enum cxobj_type type);
+cxobj    *xml_new_body(char *name, cxobj *parent, char *val);
 yang_stmt *xml_spec(cxobj *x);
 int       xml_spec_set(cxobj *x, yang_stmt *spec);
 cg_var   *xml_cv(cxobj *x);